The hello world sample of Deep Learning
Dear Ally, Before you start to enter the world of neural networks, I suggest you look at this hello world in the neural network. This is the simplest example and structure of using "Keras package" in deep learning, but all the deep learning programs, The veins cannot escape the context of this program.