The "hello world" example of Deep Learning
Dear Ally, Before you start to enter the world of neural networks, I suggest you look at this hello world in the neural network. This is the simplest example and structure of using the "Keras package" in deep learning, but all the deep learning programs, The setup of DL cannot avoid the context of this program.